Update README.md
Calculator

Define four functions - add, subtract, multiply, and divide that add/subtract/multiply/divide two numbers and print the result. Each should have two integer parameter variables.

The user is asked to input two numbers. These numbers will be passed as arguments into one of the functions.

The user is asked to input 1 to add, 2 to subtract, 3 to multiply, or 4 to divide.

If they input 1, call the 'add' method, input 2 calls the 'subtract' method, input 3 call the 'multiply' method, or input 4 call the 'divide' method.

Output the result as part of a sentence.

Sample program run:

Enter a number: 6
Enter another number: 3
Enter 1 to add, 2 to subtract, 3 to multiply, or 4 to divide: 4
6 divided by 3 is 2

Optional features you can add:

Add input validation so that a number besides 1-4 cannot be used by the user as an input when choosing the math operation.
After displaying the results of the calculations, the program should ask the user if they would like to do another calculation and exit if they do not choose yes.